+++ tests/testthat/test-get_ifdata_registry.R
@@ -1,0 +1,4 @@
+data <- get_ifdata_registry(
+  year = 2024,
+  month = 12
+)